configuring clan page with merges
